requirements.txt
face_recognition==1.3.0
opencv-python==4.9.0.80
pyzbar==0.1.9
qrcode==7.4.2
Pillow==10.2.0
numpy==1.26.4
mediapipe==0.10.9
supabase-py==2.3.4
python-dotenv==1.0.1 
tabulate